pyproject.toml
[project]
name = "afkj-data-scraper"
version = "0.1.0"
description = "Data scraper for AFK Journey"
requires-python = ">=3.11"
dependencies = [
    "pyautogui",
    "opencv-python-headless",
    "pillow",
    "easyocr",
    "gspread",
    "google-auth",
    "python-dotenv",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
]

[tool.setuptools]
packages = ["extraction", "navigation", "storage"]
py-modules = ["config", "models", "validator"]

[tool.ruff.format]
quote-style = "double"
